rubocop -A

+++ app/controllers/bookmarks_controller.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class BookmarksController < ApplicationController
   before_action :authenticate_user!
   def create
@@ -5,21 +7,13 @@ class BookmarksController < ApplicationController
     return unless @article.status == 'published'
 
     bookmark = @article.bookmarks.new(user_id: current_user.id)
-    if bookmark.save
-      redirect_to request.referer
-    else
-      redirect_to request.referer
-    end
+    redirect_to request.referer
   end
 
   def destroy
     @article = Article.find(params[:article_id])
     bookmark = @article.bookmarks.find_by(user_id: current_user.id)
-    if bookmark.present?
-      bookmark.destroy
-      redirect_to request.referer
-    else
-      redirect_to request.referer
-    end
+    bookmark.destroy if bookmark.present?
+    redirect_to request.referer
   end
 end
